+++ frontend/src/api/types.gen.ts
@@ -2862,6 +2862,11 @@ export interface components {
              * @default
              */
             price: string;
+            /**
+             * Slug
+             * @default
+             */
+            slug: string;
             /** Start Datetime */
             start_datetime?: string | null;
             /**
@@ -3050,6 +3055,11 @@ export interface components {
              * @default false
              */
             rsvp_enabled: boolean;
+            /**
+             * Slug
+             * @default
+             */
+            slug: string;
             /** Start Datetime */
             start_datetime?: string | null;
             /**
@@ -7486,6 +7496,15 @@ export interface operations {
                 };
                 content: {
                     "application/json": components["schemas"]["EventPollOut"];
+                };
+            };
+            /** @description Forbidden */
+            403: {
+                headers: {
+                    [name: string]: unknown;
+                };
+                content: {
+                    "application/json": components["schemas"]["ErrorOut"];
                 };
             };
             /** @description Not Found */